+++ frontend/app/index.js
@@ -1,20 +1,17 @@
 import React from 'react';
-import { View, Text, SafeAreaView, ScrollView, TouchableOpacity } from 'react-native';
-import { Image } from 'react-native';
+import Home from './Home';
+import MainPage from './mainpage';
+import { createStackNavigator } from '@react-navigation/stack';
 
-const Home = () => {
+const Stack = createStackNavigator();
+
+const App = () => {
   return (
-    <SafeAreaView style={{ flex: 1, backgroundColor: "#D9EDE8" }}>
-      <ScrollView contentContainerStyle={{ flexGrow: 1 }}>
-        {/* <View style={{ flex: 1, paddingTop: '10%', display: "flex", flexDirection: "column", justifyContent: "center" }}> */}
-          <View style={{ display: "flex", flexDirection: "column", justifyContent: "center", alignItems: "center",marginTop:"20%"}}>
-            <Image style={{ width: "150%" }} source={require('../main.png')} />
-            <TouchableOpacity style={{ backgroundColor: "#323D60", width: "70%", height: "13%", justifyContent: "center", alignItems: "center",marginTop:"10%",borderRadius:50 }}><Text style={{ color: "white" }}>Begin your food journey </Text></TouchableOpacity>
-          {/* </View> */}
-        </View>
-      </ScrollView>
-    </SafeAreaView>
+    <Stack.Navigator initialRouteName="Home">
+      <Stack.Screen name="Home" component={Home} />
+      <Stack.Screen name="MainPage" component={MainPage} />
+    </Stack.Navigator>
   );
 };
 
-export default Home;
+export default App;
