+++ frontend/app/mainpage.js
@@ -1,0 +1,11 @@
+import React from "react";
+import { View,Text } from "react-native";
+const MainPage =()=>{
+    return(
+        <View>
+        <Text>hello</Text>
+    </View>
+    )
+   
+}
+export default MainPage
