+++ frontend/package.json
@@ -6,6 +6,9 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.7",
+    "@react-navigation/native-stack": "^6.9.13",
+    "@react-navigation/stack": "^6.3.17",
     "axios": "^1.4.0",
     "expo": "^48.0.7",
     "expo-constants": "~14.2.1",
@@ -20,7 +23,7 @@
     "react-native-dotenv": "^3.4.9",
     "react-native-gesture-handler": "~2.9.0",
     "react-native-reanimated": "~2.14.4",
-    "react-native-safe-area-context": "4.5.0",
+    "react-native-safe-area-context": "^4.5.0",
     "react-native-screens": "~3.20.0",
     "react-native-web": "~0.18.7"
   },
